+++ src/lib.rs
@@ -250,6 +250,7 @@ mod alloc {
     pub use std::{
         borrow::{Cow, ToOwned},
         boxed::Box,
+        format,
         string::{String, ToString},
         vec::Vec,
     };
@@ -264,7 +265,7 @@ pub mod prelude {
 pub use crate::{
     claims::{Claims, Empty, TimeOptions},
     error::{Claim, CreationError, ParseError, ValidationError},
-    token::{Header, SignedToken, Token, UntrustedToken},
+    token::{Header, SignedToken, Thumbprint, Token, UntrustedToken},
     traits::{Algorithm, AlgorithmExt, AlgorithmSignature, Renamed, Validator},
 };